+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/NativeImageGenerator.java
@@ -30,6 +30,7 @@ import java.io.IOException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
+import java.nio.file.AccessDeniedException;
 import java.nio.file.FileSystems;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
@@ -54,7 +55,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-import com.oracle.svm.core.c.libc.LibCBase;
 import org.graalvm.collections.EconomicSet;
 import org.graalvm.collections.Pair;
 import org.graalvm.compiler.api.replacements.Fold;
@@ -155,8 +155,9 @@ import com.oracle.svm.core.SubstrateTargetDescription;
 import com.oracle.svm.core.SubstrateUtil;
 import com.oracle.svm.core.c.function.CEntryPointOptions;
 import com.oracle.svm.core.c.libc.GLibc;
-import com.oracle.svm.core.c.libc.MuslLibc;
+import com.oracle.svm.core.c.libc.LibCBase;
 import com.oracle.svm.core.c.libc.Libc;
+import com.oracle.svm.core.c.libc.MuslLibc;
 import com.oracle.svm.core.code.RuntimeCodeCache;
 import com.oracle.svm.core.config.ConfigurationValues;
 import com.oracle.svm.core.graal.GraalConfiguration;
@@ -443,22 +444,15 @@ public class NativeImageGenerator {
                     ImageSingletons.add(RuntimeOptionValues.class, new RuntimeOptionValues(optionProvider.getRuntimeValues(), allOptionNames));
 
                     doRun(entryPoints, javaMainSupport, imageName, k, harnessSubstitutions, compilationExecutor, analysisExecutor);
-                } finally {
+                } catch (Throwable t) {
                     try {
-                        /*
-                         * Make sure we clean up after ourselves even in the case of an exception.
-                         */
-                        if (deleteTempDirectory) {
-                            deleteAll(tempDirectory());
-                        }
-                        featureHandler.forEachFeature(Feature::cleanup);
-                    } catch (Throwable e) {
-                        /*
-                         * Suppress subsequent errors so that we unwind the original error brought
-                         * us here.
-                         */
+                        cleanup();
+                    } catch (Throwable ecleanup) {
+                        t.addSuppressed(ecleanup);
                     }
+                    throw t;
                 }
+                cleanup();
             }).get();
         } catch (InterruptedException | CancellationException e) {
             System.out.println("Interrupted!");
@@ -472,6 +466,11 @@ public class NativeImageGenerator {
         } finally {
             shutdownPoolSafe();
         }
+    }
+
+    private void cleanup() {
+        deleteTempDirectory();
+        featureHandler.forEachFeature(Feature::cleanup);
     }
 
     protected static void setSystemPropertiesForImageEarly() {
@@ -1653,7 +1652,7 @@ public class NativeImageGenerator {
     }
 
     private Path tempDirectory;
-    private boolean deleteTempDirectory;
+    private volatile boolean deleteTempDirectory;
 
     public synchronized Path tempDirectory() {
         if (tempDirectory == null) {
@@ -1674,19 +1673,34 @@ public class NativeImageGenerator {
         return tempDirectory.toAbsolutePath();
     }
 
-    private static void deleteAll(Path path) {
+    private void deleteTempDirectory() {
+        if (!deleteTempDirectory) {
+            return;
+        }
         try {
-            Files.walkFileTree(path, new SimpleFileVisitor<Path>() {
+            Files.walkFileTree(tempDirectory(), new SimpleFileVisitor<Path>() {
                 @Override
                 public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                    Files.delete(file);
+                    deleteImpl(file);
                     return FileVisitResult.CONTINUE;
                 }
 
                 @Override
                 public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                    Files.delete(dir);
+                    if (Files.isDirectory(dir) && !Files.list(dir).findAny().isPresent()) {
+                        // only delete dir if it is empty.
+                        // dirs might not be empty if a file deletion got an access denied error.
+                        deleteImpl(dir);
+                    }
                     return FileVisitResult.CONTINUE;
+                }
+
+                private void deleteImpl(Path file) throws IOException {
+                    try {
+                        Files.delete(file);
+                    } catch (AccessDeniedException e) {
+                        // we cannot do anything about access denied
+                    }
                 }
             });
         } catch (IOException ex) {