+++ truffle/CHANGELOG.md
@@ -7,6 +7,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added ability to load external default exports for libraries using a service provider. See `GenerateLibrary(defaultExportLookupEnabled = true)`.
 * The use of `@NodeField` is now permitted in combination with `@GenerateUncached`, but it throws UnsupportedOperationException when it is used.
 * It is now possible to specify a setter with `@NodeField`. The generated field then will be mutable.
+* Removed deprecated interoperability APIs that were deprecated in 19.0.0. 
+* Removed deprecated instrumentation APIs that were deprecated in 0.33.
 
 ## Version 20.0.0
 * Add [Layout#dispatch()](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/object/dsl/Layout.html#dispatch--) to be able to generate override of `ObjectType#dispatch()` method in the generated inner \*Type class.